activation_server/requirements.txt
Flask==2.3.3
cryptography==41.0.7
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
